progress

+++ db/migrate/20170516003818_create_discounts.rb
@@ -2,7 +2,7 @@ class CreateDiscounts < ActiveRecord::Migration
   def change
     create_table :discounts do |t|
       t.string :name, null: false
-      t.integer :value, null: false
+      t.decimal :value, null: false
 
       t.timestamps
     end
